# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/system": {
      "get": {
        "operationId": "AppController_getSystemInfo",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get system information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemInfo"
                }
              }
            }
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/api/flows": {
      "get": {
        "operationId": "FlowController_find",
        "parameters": [
          {
            "name": "where",
            "required": false,
            "in": "query",
            "description": "JSON object: each key is an entity field; value is a primitive or `{ \"op\": \"eq\"|\"ne\"|\"gt\"|\"gte\"|\"lt\"|\"lte\"|\"like\"|\"ilike\"|\"in\"|\"contains\", \"value\": ... }`. Must be URL-encoded when sent as a single query string.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "description": "Comma-separated field names",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "relations",
            "required": false,
            "in": "query",
            "description": "Comma-separated relation paths (e.g. `myRelation`, `myRelation.myNestedRelation`).",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "Comma-separated `field:asc|desc`.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of rows",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "example": 200
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of rows to skip",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "example": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Flow"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "flows"
        ]
      },
      "post": {
        "operationId": "FlowController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Flow"
                }
              }
            }
          }
        },
        "tags": [
          "flows"
        ]
      },
      "put": {
        "operationId": "FlowController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFlowDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Flow"
                }
              }
            }
          }
        },
        "tags": [
          "flows"
        ]
      }
    },
    "/api/flows/{id}": {
      "delete": {
        "operationId": "FlowController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "flows"
        ]
      }
    },
    "/api/nodes": {
      "get": {
        "operationId": "NodeController_find",
        "parameters": [
          {
            "name": "where",
            "required": false,
            "in": "query",
            "description": "JSON object: each key is an entity field; value is a primitive or `{ \"op\": \"eq\"|\"ne\"|\"gt\"|\"gte\"|\"lt\"|\"lte\"|\"like\"|\"ilike\"|\"in\"|\"contains\", \"value\": ... }`. Must be URL-encoded when sent as a single query string.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "description": "Comma-separated field names",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "relations",
            "required": false,
            "in": "query",
            "description": "Comma-separated relation paths (e.g. `myRelation`, `myRelation.myNestedRelation`).",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "Comma-separated `field:asc|desc`.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of rows",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "example": 200
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of rows to skip",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "example": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Node"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "nodes"
        ]
      }
    },
    "/api/packages": {
      "get": {
        "operationId": "PackageController_find",
        "parameters": [
          {
            "name": "where",
            "required": false,
            "in": "query",
            "description": "JSON object: each key is an entity field; value is a primitive or `{ \"op\": \"eq\"|\"ne\"|\"gt\"|\"gte\"|\"lt\"|\"lte\"|\"like\"|\"ilike\"|\"in\"|\"contains\", \"value\": ... }`. Must be URL-encoded when sent as a single query string.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "description": "Comma-separated field names",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "relations",
            "required": false,
            "in": "query",
            "description": "Comma-separated relation paths (e.g. `myRelation`, `myRelation.myNestedRelation`).",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "Comma-separated `field:asc|desc`.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of rows",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "example": 200
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of rows to skip",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "example": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Package"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "packages"
        ]
      }
    },
    "/api/node-types": {
      "get": {
        "operationId": "NodeTypeController_find",
        "parameters": [
          {
            "name": "where",
            "required": false,
            "in": "query",
            "description": "JSON object: each key is an entity field; value is a primitive or `{ \"op\": \"eq\"|\"ne\"|\"gt\"|\"gte\"|\"lt\"|\"lte\"|\"like\"|\"ilike\"|\"in\"|\"contains\", \"value\": ... }`. Must be URL-encoded when sent as a single query string.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "description": "Comma-separated field names",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "relations",
            "required": false,
            "in": "query",
            "description": "Comma-separated relation paths (e.g. `myRelation`, `myRelation.myNestedRelation`).",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "Comma-separated `field:asc|desc`.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of rows",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "example": 200
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of rows to skip",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "example": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NodeType"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "node-types"
        ]
      }
    },
    "/api/configs": {
      "get": {
        "operationId": "ConfigController_find",
        "parameters": [
          {
            "name": "where",
            "required": false,
            "in": "query",
            "description": "JSON object: each key is an entity field; value is a primitive or `{ \"op\": \"eq\"|\"ne\"|\"gt\"|\"gte\"|\"lt\"|\"lte\"|\"like\"|\"ilike\"|\"in\"|\"contains\", \"value\": ... }`. Must be URL-encoded when sent as a single query string.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "description": "Comma-separated field names",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "relations",
            "required": false,
            "in": "query",
            "description": "Comma-separated relation paths (e.g. `myRelation`, `myRelation.myNestedRelation`).",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "Comma-separated `field:asc|desc`.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of rows",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "example": 200
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of rows to skip",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "example": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Config"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "configs"
        ]
      }
    },
    "/api/tags": {
      "get": {
        "operationId": "TagController_find",
        "parameters": [
          {
            "name": "where",
            "required": false,
            "in": "query",
            "description": "JSON object: each key is an entity field; value is a primitive or `{ \"op\": \"eq\"|\"ne\"|\"gt\"|\"gte\"|\"lt\"|\"lte\"|\"like\"|\"ilike\"|\"in\"|\"contains\", \"value\": ... }`. Must be URL-encoded when sent as a single query string.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "description": "Comma-separated field names",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "relations",
            "required": false,
            "in": "query",
            "description": "Comma-separated relation paths (e.g. `myRelation`, `myRelation.myNestedRelation`).",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "Comma-separated `field:asc|desc`.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of rows",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "example": 200
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of rows to skip",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "example": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tag"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "tags"
        ]
      },
      "post": {
        "operationId": "TagController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tag"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "tags"
        ]
      },
      "put": {
        "operationId": "TagController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTagDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tag"
                }
              }
            }
          }
        },
        "tags": [
          "tags"
        ]
      }
    },
    "/api/tags/{id}": {
      "delete": {
        "operationId": "TagController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tags"
        ]
      }
    }
  },
  "info": {
    "title": "Snipet core API",
    "description": "The Snipet API description",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "x-api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API Key"
      }
    },
    "schemas": {
      "SystemInfo": {
        "type": "object",
        "properties": {
          "version": {
            "type": "string"
          }
        },
        "required": [
          "version"
        ]
      },
      "FlowNodeRefManifest": {
        "type": "object",
        "properties": {
          "instanceId": {
            "type": "string"
          },
          "nodeId": {
            "type": "string"
          },
          "config": {
            "type": "object"
          },
          "x": {
            "type": "number"
          },
          "y": {
            "type": "number"
          }
        },
        "required": [
          "instanceId",
          "nodeId",
          "x",
          "y"
        ]
      },
      "FlowConnectionOutManifest": {
        "type": "object",
        "properties": {
          "instanceId": {
            "type": "string"
          },
          "outputId": {
            "type": "string"
          }
        },
        "required": [
          "instanceId",
          "outputId"
        ]
      },
      "FlowConnectionInManifest": {
        "type": "object",
        "properties": {
          "instanceId": {
            "type": "string"
          },
          "inputId": {
            "type": "string"
          }
        },
        "required": [
          "instanceId",
          "inputId"
        ]
      },
      "FlowConnectionManifest": {
        "type": "object",
        "properties": {
          "source": {
            "$ref": "#/components/schemas/FlowConnectionOutManifest"
          },
          "target": {
            "$ref": "#/components/schemas/FlowConnectionInManifest"
          },
          "active": {
            "type": "boolean"
          }
        },
        "required": [
          "source",
          "target",
          "active"
        ]
      },
      "FlowManifest": {
        "type": "object",
        "properties": {
          "nodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FlowNodeRefManifest"
            }
          },
          "connections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FlowConnectionManifest"
            }
          }
        },
        "required": [
          "nodes",
          "connections"
        ]
      },
      "Flow": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "code": {
            "$ref": "#/components/schemas/FlowManifest"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "active",
          "code",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateFlowDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "code": {
            "$ref": "#/components/schemas/FlowManifest"
          },
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
      "Tag": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "NodeTag": {
        "type": "object",
        "properties": {
          "nodeId": {
            "type": "string"
          },
          "tagId": {
            "type": "string"
          },
          "tag": {
            "$ref": "#/components/schemas/Tag"
          }
        },
        "required": [
          "nodeId",
          "tagId"
        ]
      },
      "PackageTag": {
        "type": "object",
        "properties": {
          "packageId": {
            "type": "string"
          },
          "tagId": {
            "type": "string"
          },
          "tag": {
            "$ref": "#/components/schemas/Tag"
          }
        },
        "required": [
          "packageId",
          "tagId"
        ]
      },
      "Package": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "packageId": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "docs": {
            "type": "object",
            "nullable": true
          },
          "icon": {
            "type": "object",
            "nullable": true
          },
          "author": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "packageTags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PackageTag"
            }
          }
        },
        "required": [
          "id",
          "packageId",
          "version",
          "name",
          "createdAt",
          "updatedAt",
          "packageTags"
        ]
      },
      "FieldManifest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "required": {
            "type": "boolean"
          },
          "defaultValue": {
            "type": "object"
          },
          "items": {
            "$ref": "#/components/schemas/FieldManifest"
          },
          "properties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldManifest"
            }
          },
          "encrypted": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "type",
          "description"
        ]
      },
      "NodeTypeComponentManifest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "required": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "type"
        ]
      },
      "NodeTypeTag": {
        "type": "object",
        "properties": {
          "nodeTypeId": {
            "type": "string"
          },
          "tagId": {
            "type": "string"
          },
          "tag": {
            "$ref": "#/components/schemas/Tag"
          }
        },
        "required": [
          "nodeTypeId",
          "tagId"
        ]
      },
      "NodeType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "typeId": {
            "type": "string"
          },
          "packageId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "docs": {
            "type": "object",
            "nullable": true
          },
          "icon": {
            "type": "object",
            "nullable": true
          },
          "inputs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldManifest"
            }
          },
          "outputs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldManifest"
            }
          },
          "components": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NodeTypeComponentManifest"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "nodeTypeTags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NodeTypeTag"
            }
          }
        },
        "required": [
          "id",
          "typeId",
          "packageId",
          "name",
          "inputs",
          "outputs",
          "components",
          "createdAt",
          "updatedAt",
          "nodeTypeTags"
        ]
      },
      "ConfigTag": {
        "type": "object",
        "properties": {
          "configId": {
            "type": "string"
          },
          "tagId": {
            "type": "string"
          },
          "tag": {
            "$ref": "#/components/schemas/Tag"
          }
        },
        "required": [
          "configId",
          "tagId"
        ]
      },
      "Config": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "configId": {
            "type": "string"
          },
          "packageId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "docs": {
            "type": "object",
            "nullable": true
          },
          "icon": {
            "type": "object",
            "nullable": true
          },
          "fieldManifest": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldManifest"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "configTags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConfigTag"
            }
          }
        },
        "required": [
          "id",
          "configId",
          "packageId",
          "name",
          "fieldManifest",
          "createdAt",
          "updatedAt",
          "configTags"
        ]
      },
      "Node": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "nodeId": {
            "type": "string"
          },
          "packageId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "docs": {
            "type": "object",
            "nullable": true
          },
          "icon": {
            "type": "object",
            "nullable": true
          },
          "nodeTypeId": {
            "type": "string"
          },
          "configId": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "nodeTags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NodeTag"
            }
          },
          "package": {
            "$ref": "#/components/schemas/Package"
          },
          "nodeType": {
            "$ref": "#/components/schemas/NodeType"
          },
          "config": {
            "$ref": "#/components/schemas/Config"
          }
        },
        "required": [
          "id",
          "nodeId",
          "packageId",
          "name",
          "nodeTypeId"
        ]
      },
      "UpdateTagDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      }
    }
  }
}